+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* Style override for both SelectNative and SelectCustom is now called the same: `style`
+
 ## 0.3.0 (May. 20, 2013)
 
 * First public release.